Extract db config to separate file/ directory
